# Comprehensive performance improvements