--- a/.eslintignore
+++ b/.eslintignore
@@ -1,2 +1,1 @@
 seed/static/seed/js/decorators/*
-seed/static/seed/tests/*