--- a/config/settings/common.py
+++ b/config/settings/common.py
@@ -322,7 +322,7 @@ GOOGLE_RECAPTCHA_SECRET_KEY = os.environ.get("GOOGLE_RECAPTCHA_SECRET_KEY")
 
 # Certification
 # set this for a default validity_duration
-# should be a integer representing a number of days
+# should be an integer representing a number of days
 # GREEN_ASSESSMENT_DEFAULT_VALIDITY_DURATION=5 * 365
 GREEN_ASSESSMENT_DEFAULT_VALIDITY_DURATION = None