--- a/config/settings/docker.py
+++ b/config/settings/docker.py
@@ -57,6 +57,7 @@ for loc in ENV_VARS:
 DEBUG = env_var("Debug", False)
 COMPRESS_ENABLED = True
 COMPRESS_OFFLINE = True
+COMPRESS_STORAGE = "compressor.storage.BrotliCompressorFileStorage"
 
 # Make sure to disable secure cookies and csrf when using Cloudflare
 SESSION_COOKIE_SECURE = env_var("SESSION_COOKIE_SECURE", False)