--- a/seed/docs/views.py
+++ b/seed/docs/views.py
@@ -4,7 +4,6 @@ SEED Platform (TM), Copyright (c) Alliance for Sustainable Energy, LLC, and othe
 See also https://github.com/SEED-platform/seed/blob/main/LICENSE.md
 """
 
-import locale
 import os
 import re
 from collections import namedtuple
@@ -38,7 +37,7 @@ def parse_faq_file(faq_file):
     :param faq_file: str | DirEntry
     :return: FaqItem
     """
-    with open(faq_file, encoding='utf-8') as f:
+    with open(faq_file, encoding="utf-8") as f:
         _, frontmatter, body = YAML_DOC_BOUNDARY.split(f.read(), 2)
     parsed_frontmatter = yaml.safe_load(frontmatter)
     faq_item = FaqItem(