--- a/seed/static/seed/scss/style.scss
+++ b/seed/static/seed/scss/style.scss
@@ -1,4 +1,3 @@
-@import 'fonts';
 @import 'ng-wig';
 
 @keyframes fadein {